% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAD-package.R
\docType{package}
\name{LAD-package}
\alias{LAD}
\alias{LAD-package}
\title{LAD: Derive Leaf Angle Distribution (LAD) from Measured Leaf Inclination Angles}
\description{
Calculate mean statistics and leaf angle distribution type from measured leaf inclination angles. LAD distribution is fitted using a two-parameters (mu, nu) Beta distribution and compared with six theoretical LAD distributions. Additional information is provided in Chianucci and Cesaretti (2022) \doi{10.1101/2022.10.28.513998}.
}
\author{
\strong{Maintainer}: Francesco Chianucci \email{fchianucci@gmail.com} (\href{https://orcid.org/0000-0002-5688-2060}{ORCID})

Authors:
\itemize{
  \item Lorenzo Cesaretti \email{lorenzo.cesaretti@crea.gov.it}
}

}
\keyword{internal}
